tests/requirements.txt
docker==5.0.0
pytest==6.2.2
PyYAML==5.4.1